fix(model-selector-v2): fix navigation and streamline component lab playground
- replace NavigationService/window.navigate with useNavigate in model selector entry points
- prevent pin/unpin from retriggering focus scroll and forward scroll events from the virtual list container
- extract the component lab model selector playground into a dedicated content page
- move the page-level layout and tabs to ComponentLabSettings for future playground tabs
